fix(lint): remove unnecessary type assertions flagged by typescript-eslint 8.59
The dev-deps bump in #609 (typescript-eslint 8.59.x) tightened
@typescript-eslint/no-unnecessary-type-assertion. eslint --fix removed
72+ casts across 121 files; the leftover unused-import lint errors and
two type-system regressions are addressed here.

Manual fixes:
- apps/api/src/trpc/context.ts: introduce OptionalAuthEnv (auth?:
  AuthContext) in apps/api/src/lib/auth-context.ts and cast the tRPC
  Context to it. This lets c.get("auth") legitimately return undefined,
  which is what happens at runtime when no Bearer token is present —
  the previous (c.get("auth") as AuthContext | undefined) ?? null was
  flagged as unnecessary because the inline assertion looked redundant
  to the rule.
- packages/sync/src/__tests__/document-factory.test.ts: restore the
  Doc<Record<string, unknown>> annotation so Automerge.save accepts the
  union return type from createDocument(type).
- apps/api/src/__tests__/lib/bucket-access.integration.test.ts: switch
  "member" as BucketContentEntityType to "member" as const to keep
  drizzle's typed insert happy without an `as` cast.

The remaining changes are eslint --fix output: dropping `as Type` where
the receiver already accepts the source type, plus removing the now-
unused type imports those casts referenced.

+++ apps/mobile/src/platform/drivers/indexeddb-storage-adapter.ts
@@ -58,9 +58,9 @@ function snapshotToRecord(
     documentId,
     snapshotVersion: snapshot.snapshotVersion,
     ciphertext: snapshot.ciphertext,
-    nonce: snapshot.nonce as Uint8Array,
-    signature: snapshot.signature as Uint8Array,
-    authorPublicKey: snapshot.authorPublicKey as Uint8Array,
+    nonce: snapshot.nonce,
+    signature: snapshot.signature,
+    authorPublicKey: snapshot.authorPublicKey,
   };
 }
 
@@ -69,9 +69,9 @@ function changeToRecord(documentId: SyncDocumentId, change: EncryptedChangeEnvel
     documentId,
     seq: change.seq,
     ciphertext: change.ciphertext,
-    nonce: change.nonce as Uint8Array,
-    signature: change.signature as Uint8Array,
-    authorPublicKey: change.authorPublicKey as Uint8Array,
+    nonce: change.nonce,
+    signature: change.signature,
+    authorPublicKey: change.authorPublicKey,
   };
 }
